remove subresource integrity use
It was somehow breaking when used with Cloudflare -- it looks like CF is compressing the resource files but not updating the integrity hash. Also, since all the files are stored in the same place, the SRI doesn't really serve any purpose -- if the resources can be hacked, so can the HTML file.